Replaced promtheus_client with datadog statsd client + statsd-exporter
promtheus_client could not handle the use scenario under uwsgi + flask model. Its multiprocess function does not work. If not using multiprocess mode, the metrics we scape only stand for one single process after uwsgi. It seems to be reasonable to change to the latter because it reduce the complexity and is much reliable.